remove GeometryTree
