+++ server/src/test/java/org/elasticsearch/common/geo/GeoTestUtils.java
@@ -34,7 +34,7 @@ import static org.hamcrest.Matchers.equalTo;
 
 public class GeoTestUtils {
 
-    public static void assertRelation(GeoRelation expectedRelation, ShapeTreeReader reader, Extent extent) throws IOException {
+    public static void assertRelation(GeoRelation expectedRelation, TriangleTreeReader reader, Extent extent) throws IOException {
         GeoRelation actualRelation = reader.relate(extent.minX(), extent.minY(), extent.maxX(), extent.maxY());
         assertThat(actualRelation, equalTo(expectedRelation));
     }
