--- server/src/test/java/org/elasticsearch/common/geo/GeometryTreeTests.java
@@ -1,38 +1,0 @@
-/*
- * Licensed to Elasticsearch under one or more contributor
- * license agreements. See the NOTICE file distributed with
- * this work for additional information regarding copyright
- * ownership. Elasticsearch licenses this file to you under
- * the Apache License, Version 2.0 (the "License"); you may
- * not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing,
- * software distributed under the License is distributed on an
- * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
- * KIND, either express or implied.  See the License for the
- * specific language governing permissions and limitations
- * under the License.
- */
-package org.elasticsearch.common.geo;
-
-import org.elasticsearch.common.io.stream.BytesStreamOutput;
-import org.elasticsearch.geometry.Geometry;
-
-import java.io.IOException;
-
-
-public class GeometryTreeTests extends AbstractTreeTestCase {
-
-    protected ShapeTreeReader geometryTreeReader(Geometry geometry, CoordinateEncoder encoder) throws IOException {
-        GeometryTreeWriter writer = new GeometryTreeWriter(geometry, encoder);
-        BytesStreamOutput output = new BytesStreamOutput();
-        writer.writeTo(output);
-        output.close();
-        GeometryTreeReader reader = new GeometryTreeReader(encoder);
-        reader.reset(output.bytes().toBytesRef());
-        return reader;
-    }
-}
